# applications/v4/api/openapi.yaml
openapi: 3.0.3
info:
  title: Weekly Planning API
  version: 1.0.0
  description: API contract for accounts, persons, projects, weeks, and weekly planning.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
  /auth/sessions:
    post:
      tags: [Authentication]
      operationId: createAuthSession
      security: []
      requestBody:
        $ref: '#/components/requestBodies/LoginRequestBody'
      responses:
        '200': { $ref: '#/components/responses/AuthSessionCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /auth/sessions/current:
    delete:
      tags: [Authentication]
      operationId: deleteCurrentAuthSession
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
  /password-changes:
    post:
      tags: [Accounts]
      operationId: changePassword
      security: []
      requestBody:
        $ref: '#/components/requestBodies/PasswordChangeRequestBody'
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
  /weeks/{weekId}:
    get:
      tags: [Weeks]
      operationId: getWeekById
      parameters:
        - $ref: '#/components/parameters/WeekIdPathParam'
      responses:
        '200': { $ref: '#/components/responses/WeekDetails' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    put:
      tags: [Weeks]
      operationId: updateWeekById
      description: ADMIN-only. When a week is set to COMPLETED, planning write operations are rejected.
      parameters:
        - $ref: '#/components/parameters/WeekIdPathParam'
      requestBody:
        $ref: '#/components/requestBodies/UpdateWeekRequestBody'
      responses:
        '200': { $ref: '#/components/responses/WeekDetails' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    delete:
      tags: [Weeks]
      operationId: deleteWeekById
      parameters:
        - $ref: '#/components/parameters/WeekIdPathParam'
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /weeks/{weekId}/planning:
    get:
      tags: [Planning]
      operationId: getWeekPlanningBoard
      description: VIEWER users only receive planning cards linked to their own person.
      parameters:
        - $ref: '#/components/parameters/WeekIdPathParam'
        - $ref: '#/components/parameters/PlanningPersonIdQueryParam'
        - $ref: '#/components/parameters/IncludeEmptyDaysQueryParam'
      responses:
        '200': { $ref: '#/components/responses/WeekPlanningBoardResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /weeks/{weekId}/day-plans:
    get:
      tags: [Planning]
      operationId: listWeekDayPlans
      parameters:
        - $ref: '#/components/parameters/WeekIdPathParam'
        - $ref: '#/components/parameters/IncludeAssignmentsQueryParam'
      responses:
        '200': { $ref: '#/components/responses/DayPlanListResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /weeks/{weekId}/day-plans/{dayPlanId}:
    get:
      tags: [Planning]
      operationId: getWeekDayPlanById
      parameters:
        - $ref: '#/components/parameters/WeekIdPathParam'
        - $ref: '#/components/parameters/DayPlanIdPathParam'
      responses:
        '200': { $ref: '#/components/responses/DayPlanDetailsResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
...
    WeekIdPathParam:
      name: weekId
      in: path
      required: true
      schema: { type: string, format: uuid }
    DayPlanIdPathParam:
      name: dayPlanId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    PlanningPersonIdQueryParam:
      name: personId
      in: query
      schema: { type: string, format: uuid }
    IncludeEmptyDaysQueryParam:
      name: includeEmptyDays
      in: query
      schema: { type: boolean, default: true }
    IncludeAssignmentsQueryParam:
      name: includeAssignments
      in: query
      schema: { type: boolean, default: true }
  requestBodies:
    LoginRequestBody:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/LoginRequest' }
...
    PasswordChangeRequestBody:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/PasswordChangeRequest' }
...
    UpdateWeekRequestBody:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/UpdateWeekRequest' }
...
  responses:
    NoContent:
      description: No content.
    AuthSessionCreated:
      description: Authenticated session created.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/LoginResponse' }
...
    WeekDetails:
      description: Week details.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/WeekDetail' }
...
    WeekPlanningBoardResponse:
      description: Planning board.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/WeekPlanningBoard' }
    DayPlanListResponse:
      description: Week day plans.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/DayPlanCollection' }
    DayPlanDetailsResponse:
      description: Day plan details.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/DayPlan' }
...
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Unauthorized.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Forbidden.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Conflict.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    ValidationError:
      description: Validation error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    AccountRole:
      type: string
      enum: [ADMIN, VIEWER]
    PersonStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
    ProjectStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
    WeekStatus:
      type: string
      enum: [PLANNED, COMPLETED]
    DayOfWeek:
      type: string
      enum: [MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY]
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 128
          description: Username or email.
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
    LoginResponse:
      type: object
      required: [accessToken, tokenType, expiresIn, account]
      properties:
        accessToken:
          type: string
          minLength: 1
        tokenType:
          type: string
          enum: [Bearer]
        expiresIn:
          type: integer
          format: int64
          minimum: 1
        account:
          $ref: '#/components/schemas/AuthenticatedAccount'
    AuthenticatedAccount:
      type: object
      required: [accountId, personId, username, name, email, role, status]
      properties:
        accountId:
          type: string
          format: uuid
        personId:
          type: string
          format: uuid
        username:
          type: string
          maxLength: 128
        name:
          type: string
          maxLength: 120
        email:
          type: string
          format: email
          maxLength: 320
        role:
          $ref: '#/components/schemas/AccountRole'
        status:
          $ref: '#/components/schemas/PersonStatus'
...
    PasswordChangeRequest:
      type: object
      required: [username, currentPassword, newPassword, confirmNewPassword]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 128
        currentPassword:
          type: string
          format: password
          minLength: 8
          maxLength: 128
        newPassword:
          type: string
          format: password
          minLength: 8
          maxLength: 128
        confirmNewPassword:
          type: string
          format: password
          minLength: 8
          maxLength: 128
...
    UpdateWeekRequest:
      type: object
      required: [weekStart, weekEnd, status]
      properties:
        weekStart:
          type: string
          format: date
        weekEnd:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/WeekStatus'
    WeekSummary:
      type: object
      required: [weekId, weekStart, weekEnd, status]
      properties:
        weekId:
          type: string
          format: uuid
        weekStart:
          type: string
          format: date
        weekEnd:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/WeekStatus'
    WeekDetail:
      allOf:
        - $ref: '#/components/schemas/WeekSummary'
        - type: object
          required: [createdAt, updatedAt]
          properties:
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...
    WeekPlanningBoard:
      type: object
      required: [week, dayPlans]
      properties:
        week:
          $ref: '#/components/schemas/WeekDetail'
        dayPlans:
          type: array
          items:
            $ref: '#/components/schemas/DayPlan'
    DayPlanCollection:
      type: object
      required: [weekId, dayPlans]
      properties:
        weekId:
          type: string
          format: uuid
        dayPlans:
          type: array
          items:
            $ref: '#/components/schemas/DayPlan'
    DayPlan:
      type: object
      required: [dayPlanId, weekId, date, dayOfWeek, dayPersons]
      properties:
        dayPlanId:
          type: string
          format: uuid
        weekId:
          type: string
          format: uuid
        date:
          type: string
          format: date
        dayOfWeek:
          $ref: '#/components/schemas/DayOfWeek'
        dayPersons:
          type: array
          items:
            $ref: '#/components/schemas/DayPerson'
    DayPerson:
      type: object
      required: [dayPersonId, person, dayPersonProjects]
      properties:
        dayPersonId:
          type: string
          format: uuid
        person:
          $ref: '#/components/schemas/PersonReference'
        dayPersonProjects:
          type: array
          items:
            $ref: '#/components/schemas/DayPersonProject'
    DayPersonProject:
      type: object
      required: [dayPersonProjectId, project, tasks]
      properties:
        dayPersonProjectId:
          type: string
          format: uuid
        project:
          $ref: '#/components/schemas/ProjectReference'
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
    Task:
      type: object
      required: [taskId, description, createdAt, updatedAt]
      properties:
        taskId:
          type: string
          format: uuid
        description:
          type: string
          minLength: 1
          maxLength: 2000
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PersonReference:
      type: object
      required: [personId, name, email, status]
      properties:
        personId:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 120
        email:
          type: string
          format: email
          maxLength: 320
        status:
          $ref: '#/components/schemas/PersonStatus'
    ProjectReference:
      type: object
      required: [projectId, name, code, status]
      properties:
        projectId:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 120
        code:
          type: string
          maxLength: 40
        status:
          $ref: '#/components/schemas/ProjectStatus'
...
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
        rejectedValue:
          type: string
          nullable: true
    ErrorResponse:
      type: object
      required: [timestamp, status, error, code, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        code:
          type: string
        message:
          type: string
        path:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'